+++ trace/config_test.go
@@ -70,7 +70,7 @@ func TestNewSpanConfig(t *testing.T) {
 		},
 		{
 			[]SpanStartOption{
-				WithLazyAttributes(func() []attribute.KeyValue { return []attribute.KeyValue{k1v1} }),
+				WithAttributesLazy(func() []attribute.KeyValue { return []attribute.KeyValue{k1v1} }),
 			},
 			SpanConfig{
 				attributes: []attribute.KeyValue{k1v1},
@@ -79,9 +79,9 @@ func TestNewSpanConfig(t *testing.T) {
 		{
 			// Multiple calls should append not overwrite.
 			[]SpanStartOption{
-				WithLazyAttributes(func() []attribute.KeyValue { return []attribute.KeyValue{k1v1} }),
-				WithLazyAttributes(func() []attribute.KeyValue { return []attribute.KeyValue{k1v2} }),
-				WithLazyAttributes(func() []attribute.KeyValue { return []attribute.KeyValue{k2v2} }),
+				WithAttributesLazy(func() []attribute.KeyValue { return []attribute.KeyValue{k1v1} }),
+				WithAttributesLazy(func() []attribute.KeyValue { return []attribute.KeyValue{k1v2} }),
+				WithAttributesLazy(func() []attribute.KeyValue { return []attribute.KeyValue{k2v2} }),
 			},
 			SpanConfig{
 				// No uniqueness is guaranteed by the API.
@@ -90,7 +90,7 @@ func TestNewSpanConfig(t *testing.T) {
 		},
 		{
 			[]SpanStartOption{
-				WithLazyAttributes(func() []attribute.KeyValue { return []attribute.KeyValue{k1v1, k1v2, k2v2} }),
+				WithAttributesLazy(func() []attribute.KeyValue { return []attribute.KeyValue{k1v1, k1v2, k2v2} }),
 			},
 			SpanConfig{
 				// No uniqueness is guaranteed by the API.
@@ -210,7 +210,7 @@ func TestSpanStartConfigLazyAttributeMutability(t *testing.T) {
 	a := attribute.String("a", "val")
 	b := attribute.String("b", "val")
 	attrs := []attribute.KeyValue{a, b}
-	conf := NewSpanStartConfig(WithLazyAttributes(func() []attribute.KeyValue {
+	conf := NewSpanStartConfig(WithAttributesLazy(func() []attribute.KeyValue {
 		return attrs
 	}))
 
@@ -223,7 +223,7 @@ func TestSpanStartConfigLazyAttributeMutability(t *testing.T) {
 
 func TestSpanStartConfigLazyAttributeMultipleCalls(t *testing.T) {
 	callCount := int64(0)
-	lazyAttrs := WithLazyAttributes(func() []attribute.KeyValue {
+	lazyAttrs := WithAttributesLazy(func() []attribute.KeyValue {
 		return []attribute.KeyValue{
 			attribute.Int64("a", atomic.AddInt64(&callCount, 1)),
 		}